Fermat Core cleaning part 3 | Compiling & Running
