--- DMP/android/reference_niche_wallet/fermat-dmp-android-reference-niche-wallet-bitcoin-wallet-bitdubai/src/main/res/layout/fragment_database_tools.xml
@@ -1,32 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:orientation="vertical"
-    android:background="#000000"
-    android:id="@+id/hola">
-
-    <TextView
-        android:id="@+id/labelDatabase"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:gravity="center"
-        android:paddingLeft="7dip"
-        android:paddingRight="7dip"
-        android:paddingTop="3dip"
-        android:paddingBottom="3dip"
-        android:textColor="#FF666666"
-        android:textSize="30sp"
-        android:text="Click on a Resource"/>
-
-    <ListView
-        android:id="@+id/lista1"
-        android:layout_height="wrap_content"
-        android:layout_width="match_parent"
-        android:layout_weight="0.14"
-        android:headerDividersEnabled="false"
-        android:divider="#FF0000"
-        android:textColor="#000000">
-    </ListView>
-</LinearLayout>
