+++ PIP/android/sub_app/fermat-pip-android-sub-app-developer-bitdubai/src/main/java/com/bitdubai/sub_app/developer/fragment/DatabaseToolsDatabaseTableListFragment.java
@@ -16,10 +16,11 @@ import android.widget.TextView;
 import com.bitdubai.sub_app.developer.R;
 import com.bitdubai.fermat_api.layer.all_definition.developer.DeveloperDatabase;
 import com.bitdubai.fermat_api.layer.all_definition.developer.DeveloperDatabaseTable;
+import com.bitdubai.fermat_api.layer.all_definition.enums.Addons;
+import com.bitdubai.fermat_api.layer.all_definition.enums.Plugins;
 import com.bitdubai.fermat_api.layer.pip_actor.developer.DatabaseTool;
-import com.bitdubai.sub_app.developer.fragment.Platform;
+import com.bitdubai.fermat_api.layer.pip_actor.developer.ToolManager;
 
-import java.util.ArrayList;
 import java.util.List;
 
 /**
@@ -36,13 +37,11 @@ public class DatabaseToolsDatabaseTableListFragment extends Fragment {
     private static final String ARG_POSITION = "position";
     View rootView;
 
+    private String resource;
+
     private DatabaseTool databaseTools;
 
     private DeveloperDatabase developerDatabase;
-
-    public void setDeveloperDatabase(DeveloperDatabase developerDatabase) {
-        this.developerDatabase = developerDatabase;
-    }
 
     List<DeveloperDatabaseTable> developerDatabaseTableList;
 
@@ -60,7 +59,6 @@ public class DatabaseToolsDatabaseTableListFragment extends Fragment {
     public void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
 
-/*
         try {
             ToolManager toolManager = platform.getToolManager();
             try {
@@ -73,8 +71,6 @@ public class DatabaseToolsDatabaseTableListFragment extends Fragment {
             showMessage("Unexpected error get Transactions - " + ex.getMessage());
             ex.printStackTrace();
         }
-
-        */
     }
 
     @Override
@@ -83,24 +79,21 @@ public class DatabaseToolsDatabaseTableListFragment extends Fragment {
         rootView = inflater.inflate(R.layout.fragment_database_tools, container, false);
 
         try {
-            //developerDatabaseTableList = databaseTools.getTableListFromDatabase(developerDatabase);
-        } catch (Exception e) {
-            System.out.println("***********************hasta la vaina, baby");
-        }
+            String name = resource.split(" - ")[0];
+            String type = resource.split(" - ")[1];
+            if (type.equals("Addon")) {
+                Addons addon = Addons.getByKey(name);
+                this.developerDatabaseTableList = databaseTools.getAddonTableListFromDatabase(addon, developerDatabase);
+            } else if (type.equals("Plugin")) {
+                Plugins plugin = Plugins.getByKey(name);
+                this.developerDatabaseTableList = databaseTools.getPluginTableListFromDatabase(plugin, developerDatabase);
+            }
 
-        try {
             // Get ListView object from xml
             final ListView listView = (ListView) rootView.findViewById(R.id.lista1);
 
             TextView labelDatabase = (TextView) rootView.findViewById(R.id.labelDatabase);
-            labelDatabase.setText("Click on a Table");
-
-            developerDatabaseTableList = new ArrayList<>();
-
-            developerDatabaseTableList.add(new DeveloperDatabaseTableTest("incoming_crypto_transaction", null));
-            developerDatabaseTableList.add(new DeveloperDatabaseTableTest("outgoing_crypto_transaction", null));
-            developerDatabaseTableList.add(new DeveloperDatabaseTableTest("crypto_transaction", null));
-            developerDatabaseTableList.add(new DeveloperDatabaseTableTest("wallet_crypto_address", null));
+            labelDatabase.setText(developerDatabase.getName()+" - Tables List");
 
             final List<DeveloperDatabaseTable> developerDatabaseTableList2 = developerDatabaseTableList;
 
@@ -121,49 +114,32 @@ public class DatabaseToolsDatabaseTableListFragment extends Fragment {
                 public void onItemClick(AdapterView<?> parent, View view, int position, long id) {
                 String item = (String) listView.getItemAtPosition(position);
 
-                    for (DeveloperDatabaseTable devDBTable : developerDatabaseTableList2) {
-                        if (devDBTable.getName().equals(item)) {
-                            DatabaseToolsDatabaseTableRecordListFragment databaseToolsDatabaseTableRecordListFragment = new DatabaseToolsDatabaseTableRecordListFragment();
-                            databaseToolsDatabaseTableRecordListFragment.setDeveloperDatabase(developerDatabase);
-                            databaseToolsDatabaseTableRecordListFragment.setDeveloperDatabaseTable(devDBTable);
+                for (DeveloperDatabaseTable devDBTable : developerDatabaseTableList2) {
+                    if (devDBTable.getName().equals(item)) {
+                        DatabaseToolsDatabaseTableRecordListFragment databaseToolsDatabaseTableRecordListFragment = new DatabaseToolsDatabaseTableRecordListFragment();
+                        databaseToolsDatabaseTableRecordListFragment.setResource(resource);
+                        databaseToolsDatabaseTableRecordListFragment.setDeveloperDatabase(developerDatabase);
+                        databaseToolsDatabaseTableRecordListFragment.setDeveloperDatabaseTable(devDBTable);
 
-                            FragmentTransaction FT = getFragmentManager().beginTransaction();
+                        FragmentTransaction FT = getFragmentManager().beginTransaction();
 
-                            FT.replace(R.id.hola, databaseToolsDatabaseTableRecordListFragment);
+                        FT.replace(R.id.hola, databaseToolsDatabaseTableRecordListFragment);
 
-                            FT.commit();
-                        }
+                        FT.addToBackStack(null);
+
+                        FT.commit();
                     }
+                }
 
                 }
             });
 
             listView.setAdapter(adapter);
         } catch (Exception e) {
-            showMessage("DatabaseTools Fragment onCreateView Exception - " + e.getMessage());
+            showMessage("DatabaseTools Database Table List Fragment onCreateView Exception - " + e.getMessage());
             e.printStackTrace();
         }
         return rootView;
-    }
-
-    public class DeveloperDatabaseTableTest implements DeveloperDatabaseTable {
-        String name;
-        List<String> fieldsNames;
-
-        DeveloperDatabaseTableTest(String name, List<String> fieldsNames) {
-            this.name = name;
-            this.fieldsNames = fieldsNames;
-        }
-
-        @Override
-        public String getName() {
-            return name;
-        }
-
-        @Override
-        public List<String> getFieldNames() {
-            return fieldsNames;
-        }
     }
 
     //show alert
@@ -178,5 +154,13 @@ public class DatabaseToolsDatabaseTableListFragment extends Fragment {
         });
         //alertDialog.setIcon(R.drawable.icon);
         alertDialog.show();
+    }
+
+    public void setDeveloperDatabase(DeveloperDatabase developerDatabase) {
+        this.developerDatabase = developerDatabase;
+    }
+
+    public void setResource(String resource) {
+        this.resource = resource;
     }
 }
