+++ PIP/android/sub_app/fermat-pip-android-sub-app-developer-bitdubai/src/main/res/layout/fragment_log_tools.xml
