+++ fermat-api/src/main/java/com/bitdubai/fermat_api/layer/dmp_middleware/app_runtime/enums/Fragments.java
@@ -18,8 +18,7 @@ public enum Fragments {
     CWP_WALLET_RUNTIME_WALLET_BITCOIN_ALL_BITDUBAI_RECEIVE,
     CWP_WALLET_RUNTIME_WALLET_BITCOIN_ALL_BITDUBAI_TRANSACTIONS,
     CWP_WALLET_RUNTIME_WALLET_BITCOIN_ALL_BITDUBAI_CONTACTS,
-    CWP_WALLET_RUNTIME_WALLET_BITCOIN_ALL_BITDUBAI_DATABASE_TOOLS,
-    CWP_WALLET_RUNTIME_WALLET_BITCOIN_ALL_BITDUBAI_LOG_TOOLS,
+
     CWP_WALLET_RUNTIME_WALLET_BITCOIN_ALL_BITDUBAI_CONTACTS_SEND,
     CWP_WALLET_RUNTIME_WALLET_BITCOIN_ALL_BITDUBAI_CONTACTS_NEW_SEND,
     CWP_WALLET_RUNTIME_WALLET_BITCOIN_ALL_BITDUBAI_CONTACTS_RECEIVE,
@@ -59,7 +58,8 @@ public enum Fragments {
     CWP_WALLET_ADULTS_ALL_DAILY_DISCOUNT,
     CWP_WALLET_ADULTS_ALL_WEEKLY_DISCOUNT,
     CWP_WALLET_ADULTS_ALL_MONTHLY_DISCOUNT,
-    CWP_SUB_APP_DEVELOPER_MAIN,
+    CWP_SUB_APP_DEVELOPER_DATABASE_TOOLS,
+    CWP_SUB_APP_DEVELOPER_LOG_TOOLS,
     CWP_WALLET_STORE_MAIN,
     CWP_WALLET_FACTORY_MAIN;
