+++ fermat-api/src/main/java/com/bitdubai/fermat_api/layer/dmp_middleware/app_runtime/enums/Wallets.java
@@ -9,5 +9,6 @@ public enum Wallets {
     CWP_WALLET_RUNTIME_WALLET_AGE_TEEN_ALL_BITDUBAI,
     CWP_WALLET_RUNTIME_WALLET_ADULTS_ALL_BITDUBAI,
     CWP_WALLET_RUNTIME_WALLET_AGE_YOUNG_ALL_BITDUBAI,
-    CWP_WALLET_RUNTIME_WALLET_BITCOIN_WALLET_ALL_BITDUBAI;
+    CWP_WALLET_RUNTIME_WALLET_BITCOIN_WALLET_ALL_BITDUBAI,
+    CWP_WALLET_RUNTIME_WALLET_DEVELOPER_APP_ALL_BITDUBAI;
 }
